refactor: static toolbar modularization
