+++ src/frontend/web/src/components/Note/EditorFrame/index.tsx
@@ -1,17 +1,8 @@
 import Caret from './Caret';
 import React, { useCallback } from 'react';
 import { Node } from 'slate';
-import {
-  Editable,
-  ReactEditor,
-  RenderLeafProps,
-  Slate,
-  useSlate,
-} from 'slate-react';
-import { ClientFrame, Icon, IconButton } from './Components';
-import { isBlockActive, toggleBlock } from './plugin/block';
-import { insertLink, isLinkActive, unwrapLink } from './plugin/link';
-import { isMarkActive, toggleMark } from './plugin/mark';
+import { Editable, ReactEditor, RenderLeafProps, Slate } from 'slate-react';
+import { ClientFrame } from './Components';
 
 export interface EditorFrame {
   editor: ReactEditor;
@@ -35,30 +26,6 @@ const EditorFrame: React.FC<EditorFrame> = ({
   return (
     <ClientFrame>
       <Slate editor={editor} value={value} onChange={onChange}>
-        <div
-          style={{
-            display: 'flex',
-            flexWrap: 'wrap',
-            position: 'sticky',
-            top: 0,
-            backgroundColor: 'white',
-            zIndex: 1,
-          }}
-        >
-          <MarkButton format="bold" icon="format_bold" />
-          <MarkButton format="italic" icon="format_italic" />
-          <MarkButton format="underline" icon="format_underlined" />
-          <MarkButton format="code" icon="code" />
-
-          <BlockButton format="heading-one" icon="looks_one" />
-          <BlockButton format="heading-two" icon="looks_two" />
-          <BlockButton format="block-quote" icon="format_quote" />
-
-          <BlockButton format="numbered-list" icon="format_list_numbered" />
-          <BlockButton format="bulleted-list" icon="format_list_bulleted" />
-
-          <LinkButton />
-        </div>
         <Editable
           readOnly={readOnly}
           renderElement={renderElement}
@@ -134,58 +101,5 @@ const Leaf: React.FC<RenderLeafProps> = ({ attributes, children, leaf }) => {
       {leaf.isCaret ? <Caret {...(leaf as any)} /> : null}
       {children}
     </span>
-  );
-};
-
-const BlockButton: React.FC<any> = ({ format, icon }) => {
-  const editor = useSlate();
-  return (
-    <IconButton
-      active={isBlockActive(editor, format)}
-      onMouseDown={(event: React.MouseEvent) => {
-        event.preventDefault();
-        toggleBlock(editor, format);
-      }}
-    >
-      <Icon className="material-icons">{icon}</Icon>
-    </IconButton>
-  );
-};
-
-const MarkButton: React.FC<any> = ({ format, icon }) => {
-  const editor = useSlate();
-  return (
-    <IconButton
-      active={isMarkActive(editor, format)}
-      onMouseDown={(event: React.MouseEvent) => {
-        event.preventDefault();
-        toggleMark(editor, format);
-      }}
-    >
-      <Icon className="material-icons">{icon}</Icon>
-    </IconButton>
-  );
-};
-
-const LinkButton = () => {
-  const editor = useSlate();
-
-  const isActive = isLinkActive(editor);
-
-  return (
-    <IconButton
-      active={isActive}
-      onMouseDown={(event: React.MouseEvent) => {
-        event.preventDefault();
-
-        if (isActive) return unwrapLink(editor);
-
-        const url = window.prompt('Enter the URL of the link:');
-
-        url && insertLink(editor, url);
-      }}
-    >
-      <Icon className="material-icons">link</Icon>
-    </IconButton>
   );
 };
